+++ CHANGELOG.rst
@@ -113,7 +113,7 @@ Unreleased
 0.1.3
 -----
 
-- Changes to allow ingest of an arbitrary catalog for testing.
+- Changes to allow ingest of an arbitrary catalogue for testing.
   (`MR22 <https://gitlab.com/ska-telescope/sdp/ska-sdp-global-sky-model/-/merge_requests/22>`__)
 - Improve Local Sky Model creation speed
   (`MR21 <https://gitlab.com/ska-telescope/sdp/ska-sdp-global-sky-model/-/merge_requests/21>`__)
@@ -125,7 +125,7 @@ Unreleased
   (`MR18 <https://gitlab.com/ska-telescope/sdp/ska-sdp-global-sky-model/-/merge_requests/18>`__)
 - Fixed up dockerfile, ensure DB connection is up and running before finishing app startup
   (`MR17 <https://gitlab.com/ska-telescope/sdp/ska-sdp-global-sky-model/-/merge_requests/17>`__)
-- Make the ingest generic to be able to process both vizier and file based catalog
+- Make the ingest generic to be able to process both vizier and file based catalogue
   (`MR16 <https://gitlab.com/ska-telescope/sdp/ska-sdp-global-sky-model/-/merge_requests/16>`__)
 - Implement Cone Search
   (`MR15 <https://gitlab.com/ska-telescope/sdp/ska-sdp-global-sky-model/-/merge_requests/15>`__)
@@ -137,7 +137,7 @@ Unreleased
   (`MR12 <https://gitlab.com/ska-telescope/sdp/ska-sdp-global-sky-model/-/merge_requests/12>`__)
 - Add probes, update volumes, update docker compose
   (`MR11 <https://gitlab.com/ska-telescope/sdp/ska-sdp-global-sky-model/-/merge_requests/11>`__)
-- Ingest GLEAM catalog, updated models, added test endpoints
+- Ingest GLEAM catalogue, updated models, added test endpoints
   (`MR10 <https://gitlab.com/ska-telescope/sdp/ska-sdp-global-sky-model/-/merge_requests/10>`__)
 - Use sqlalchemy for Data models
   (`MR7 <https://gitlab.com/ska-telescope/sdp/ska-sdp-global-sky-model/-/merge_requests/7>`__)